Show Guest List before Recent Activity on small screens
Now using a grid layout for bookings so that the attendee information
can remain on the right-hand side on larger screens whilst collapsing
down between the booking details and the attendee list on smaller
screens. This also ensures that the guest list does not push the recent
activity further down large screens when there is a lot of information
to display.